+++ jericho/plugin/threaded_async_http.py
@@ -1,4 +1,5 @@
 #!/bin/python3
+import time
 import queue
 import threading
 import asyncio
@@ -22,8 +23,11 @@ class ThreadedAsyncHTTP:
         self.parts: typing.List[typing.List[str]] = []
         self.async_http: AsyncHTTP = async_http
         self.user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36"
+        self.work_initialized = False
 
         self.finish_queue: queue.Queue = queue.Queue()
+        self.single_queue: queue.Queue = queue.Queue()
+        self.wip_queue: queue.Queue = queue.Queue()
 
         for worker_id in range(num_threads):
             thread_queue: queue.Queue = queue.Queue()
@@ -41,7 +45,7 @@ class ThreadedAsyncHTTP:
         self, domains: typing.List[str]
     ) -> typing.List[typing.Tuple]:
         """Send a HEAD requests to a list of domains"""
-        return await self.async_http.head(
+        res =  await self.async_http.head(
             domains,
             settings={
                 "status": HttpStatusCode.OK.value,
@@ -51,11 +55,14 @@ class ThreadedAsyncHTTP:
             },
         )
 
+        self.wip_queue.get()
+        return res
+
     async def _send_get_request(
         self, domains: typing.List[str]
     ) -> typing.List[typing.Tuple]:
         """Send a GET requests to a list of domains"""
-        return await self.async_http.get(
+        res = await self.async_http.get(
             domains,
             settings={
                 "status": HttpStatusCode.OK.value,
@@ -64,6 +71,9 @@ class ThreadedAsyncHTTP:
                 "headers": {"User-Agent": self.user_agent},
             },
         )
+
+        self.wip_queue.get()
+        return res
 
     def _send(self, single_queue: queue.Queue, finish_queue: queue.Queue) -> typing.Any:
         """
@@ -79,35 +89,56 @@ class ThreadedAsyncHTTP:
             else:
                 method, domains = payload
                 if method == HttpRequestMethods.HEAD:
-                    res = asyncio.run(self._send_head_request(domains))
+                    self.wip_queue.put(True)
+                    head_responsive = [
+                        domain
+                        for domain, _, _ in asyncio.run(
+                            self._send_head_request(domains)
+                        )
+                    ]
+                    single_queue.put((HttpRequestMethods.GET, head_responsive))
 
                 elif method == HttpRequestMethods.GET:
+                    self.wip_queue.put(True)
                     res = asyncio.run(self._send_get_request(domains))
+                    finish_queue.put(res)
                 else:
                     self.close()
                     raise InvalidHTTPRequestMethod
 
-                finish_queue.put(res)
+    def _is_idle(self):
+        time.sleep(.1)
+        empty_work_queues = [
+            single_queue
+            for _, single_queue in enumerate(self.queues)
+            if single_queue.empty()
+        ]
+        logging.debug(
+            f"Checking idle - Work queue: {len(empty_work_queues) == len(self.queues)} - Finish queue: {self.finish_queue.empty()} - WIP queue: {self.wip_queue.qsize()}"
+        )
+        return (
+            len(empty_work_queues) == len(self.queues)
+            and self.finish_queue.empty()
+            and self.wip_queue.empty()
+        )
 
     def get_response(self) -> typing.List[typing.Tuple]:
         """Loop through all of the work queues and combine the data"""
+        while not self._is_idle():
+            if not self.finish_queue.empty():
+                for res in self.finish_queue.get():
+                    yield res
+        return False
 
-        output: typing.List[typing.Tuple] = []
-        for _ in self.parts:
-            res = self.finish_queue.get()
-            output = output + res
-        return output
-
-    def start_bulk(
-        self, domains: typing.List[str], method: HttpRequestMethods
-    ) -> typing.Any:
+    def start_bulk(self, domains: typing.List[str]) -> typing.Any:
         """Supply domains list to all the threads"""
 
         self.parts = split_array_by(domains, len(self.threads))
-
         for key, single_queue in enumerate(self.queues):
-            if key < len(self.parts):  # The data might not be enough for 10 threads
-                single_queue.put((method, self.parts[key]))
+            if key < len(
+                self.parts
+            ):  # The data might not be enough for the amount of threads
+                single_queue.put((HttpRequestMethods.HEAD, self.parts[key]))
 
     def close(self) -> typing.Any:
         """Send a kill message to all worker threads and join the threads"""